requirements.txt
supabase==2.28.0
Pillow==10.4.0
tomli==2.0.1; python_version < "3.11"
